+++ examples/MemoryBenchmark/generate_readme.py
@@ -38,7 +38,7 @@ by the runtime environment of the processor. For example, it often seems like
 the ESP8266 allocates flash memory in blocks of a certain quantity, so the
 calculated flash size can jump around in unexpected ways.
 
-**Version**: AceSegment v0.10.0
+**Version**: AceSegment v0.11.0
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.